MAINT: ContextMenu をややリファクタリング
