+++ apps/browser/src/contents/components/ContextMenu/ContextMenu.tsx
@@ -1,32 +1,17 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-  HighlighterIcon,
-  MoreVerticalIcon,
-  PowerIcon,
-  PowerOffIcon,
-} from "lucide-react";
+import { useCallback, useEffect, useRef, useState } from "react";
 
 import type { TagDTO } from "@whl/common-types";
-import { Button } from "@whl/ui/components/ui/button";
-import { Label } from "@whl/ui/components/ui/label";
 import {
   Popover,
   PopoverAnchor,
   PopoverContent,
-  PopoverTrigger,
 } from "@whl/ui/components/ui/popover";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@whl/ui/components/ui/tooltip";
 
 import { useHighlight } from "~/contents/hooks/useLabel";
 import { useLabels } from "~/contents/hooks/useLabels";
 import { usePopover } from "~/contents/hooks/usePopover";
-import { useIgnoredDomains } from "~/hooks/useIgnoredDomain";
 import { useSession } from "~/hooks/useSession";
+import Actions from "./Actions";
 import Labels from "./Labels";
 import ShortcutHighlight from "./ShortcutHighlight";
 import TagForm from "./TagForm";
@@ -38,12 +23,6 @@ const ContextMenu = () => {
   const { status } = useSession();
   const [tags, setTags] = useState<TagDTO[]>([]);
   const anchor = useRef<HTMLDivElement>(null);
-  const [ignoredDomains, { toggle }] = useIgnoredDomains();
-
-  const enabled = useMemo<boolean>(
-    () => ignoredDomains.includes(window.location.hostname) === false,
-    [ignoredDomains],
-  );
 
   useEffect(() => {
     if (!open) {
@@ -92,6 +71,8 @@ const ContextMenu = () => {
         />
       </PopoverAnchor>
       <PopoverContent
+        // eslint-disable-next-line jsx-a11y/no-autofocus
+        autoFocus={false}
         className="whl-w-auto whl-max-w-xs whl-p-0"
         onMouseUp={(e) => {
           e.preventDefault();
@@ -99,64 +80,7 @@ const ContextMenu = () => {
         }}
       >
         {highlight === null ? (
-          <Popover>
-            <div className="whl-rounded whl-bg-primary whl-px-2 whl-py-1">
-              <div className="whl-flex whl-flex-row whl-gap-2">
-                <TooltipProvider>
-                  <Tooltip delayDuration={400}>
-                    <div className="whl-flex whl-flex-row whl-gap-2">
-                      <TooltipTrigger asChild>
-                        <Button size="icon_sm" onClick={setDefaultHighlight}>
-                          <HighlighterIcon size={24} />
-                          Mark!
-                        </Button>
-                      </TooltipTrigger>
-                      <Button size="icon_sm">
-                        <MoreVerticalIcon size={24} onClick={handleOpenMore} />
-                      </Button>
-                    </div>
-                    <TooltipContent side="bottom">
-                      <p className="whl-font-mono whl-text-xs">
-                        Highlight Text (alt+c)
-                      </p>
-                    </TooltipContent>
-                  </Tooltip>
-                </TooltipProvider>
-                <Button size="icon_sm" onClick={setDefaultHighlight}>
-                  <HighlighterIcon size={20} />
-                </Button>
-                <PopoverTrigger asChild>
-                  <Button size="icon_sm">
-                    <MoreVerticalIcon size={20} />
-                  </Button>
-                </PopoverTrigger>
-              </div>
-              <PopoverContent
-                align="start"
-                className="whl-w-auto whl-bg-primary whl-p-0"
-              >
-                <div className="whl-flex whl-flex-col whl-gap-0.5">
-                  <Button
-                    onClick={() => toggle(window.location.hostname)}
-                    className="whl-gap-1"
-                  >
-                    <Label
-                      htmlFor="enabled-on"
-                      className="whl-text-primary-foreground"
-                    >
-                      {enabled ? "Disable" : "Enable"} on{" "}
-                      {window.location.hostname}
-                    </Label>
-                    {enabled ? (
-                      <PowerOffIcon size={16} />
-                    ) : (
-                      <PowerIcon size={16} />
-                    )}
-                  </Button>
-                </div>
-              </PopoverContent>
-            </div>
-          </Popover>
+          <Actions handleHighlight={setDefaultHighlight} />
         ) : (
           <div className="whl-flex whl-flex-col whl-gap-2 whl-p-2">
             <Labels labels={labels} onChanged={setLabel} />
