+++ apps/browser/src/contents/components/contexts/HighlightsProvider.tsx
@@ -27,7 +27,6 @@ export const HighlightsProvider = ({ children }: PropsWithChildren) => {
   const { mark } = useMarker();
 
   useEffect(() => {
-    console.log("highlights", highlights);
     highlights.forEach((marker) => {
       if (marker.position) {
         mark(marker.position, marker.label.color);
